Add TypeMatchup model and tests
A TypeMatchup represents the damage multiplier corresponding to
an attack type and a defending type. I put them in db/seeds.rb as
they are a source of constant data that is needed in multiple places
in the application and it would be cumbersome to write them all in a
class file that contains other useful code.